add film folder
