+++ testing/test_sliceFilter.py
@@ -5,6 +5,7 @@ import torch
 import torch.backends.cudnn as cudnn
 from torch.utils.data import DataLoader
 from torchvision import transforms
+from medicaltorch import transforms as mt_transforms
 
 import ivadomed.transforms as ivadomed_transforms
 from ivadomed.loader import utils as imed_loader_utils, loader as imed_loader
